+++ baybe/optimizer.py
@@ -1,3 +1,0 @@
-"""
-Optimizer functionality such as Gaussian processes, random forests etc.
-"""