+++ baybe/utils.py
@@ -2,6 +2,12 @@
 Collection of small utilities
 """
 
+from typing import Iterable, Union
+
+import pandas as pd
+import torch
+from torch import Tensor
+
 
 def is_valid_smiles(smiles: str):
     """
@@ -10,3 +16,11 @@ def is_valid_smiles(smiles: str):
     :return: True if the SMILES is valid according to RDKit
     """
     return smiles
+
+
+def to_tensor(*dfs: pd.DataFrame) -> Union[Tensor, Iterable[Tensor]]:
+    """Converts a given set of dataframes into tensors (dropping all indices)."""
+    out = (torch.from_numpy(df.values).to(torch.float32) for df in dfs)
+    if len(dfs) == 1:
+        out = next(out)
+    return out